refact: move components position
